PO-9740: Legacy action codes changes and BigInteger version

+++ src/main/java/uk/gov/hmcts/opal/service/opal/OpalDefendantAccountEnforcementService.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.opal.service.opal;
 
 import tools.jackson.core.JacksonException;
 import tools.jackson.databind.ObjectMapper;
+import java.math.BigInteger;
 import java.time.Clock;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -157,7 +158,7 @@ public class OpalDefendantAccountEnforcementService
 
         return AddEnforcementResponse.builder()
             .defendantAccountId(String.valueOf(defendantAccountId))
-            .version(Math.toIntExact(latestDefendant.getVersionNumber()))
+            .version(BigInteger.valueOf(latestDefendant.getVersionNumber()))
             .enforcementId(String.valueOf(enforcementId))
             .build();
     }
